Major changes in socket data structure, compressed the size almost 3 times. Refactored how some fields are read and so on. Other types of refactoring.

+++ src/Chunks/BMX/utils.cpp
@@ -171,7 +171,7 @@ __forceinline uint32_t __stdcall sc_bmx_write_to_chunks(
         *(uint32_t*)(cur_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_USER_DATA_WRITTEN_BYTES) = buf_len_bytes;
 
         // Setting total number of chunks.
-        *(int32_t*)(cur_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_NUM_CHUNKS) = 1;
+        *(uint16_t*)(cur_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_NUM_CHUNKS) = 1;
 
         // Setting number of total written bytes.
         *actual_written_bytes = buf_len_bytes;
@@ -193,7 +193,7 @@ __forceinline uint32_t __stdcall sc_bmx_write_to_chunks(
     }
 
     // Setting total number of chunks.
-    *(int32_t*)(cur_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_NUM_CHUNKS) = 1 + num_extra_chunks_to_use;
+    *(uint16_t*)(cur_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_NUM_CHUNKS) = 1 + num_extra_chunks_to_use;
 
     // Acquiring linked chunks.
     err_code = cm_acquire_linked_shared_memory_chunks_counted(cur_chunk_index, num_extra_chunks_to_use);
@@ -276,7 +276,7 @@ __forceinline uint32_t __stdcall sc_bmx_send_small_buffer(
     *(uint32_t*)(src_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_USER_DATA_WRITTEN_BYTES) = buf_len_bytes;
 
     // Setting total number of chunks.
-    *(int32_t*)(src_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_NUM_CHUNKS) = 1;
+    *(uint16_t*)(src_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_NUM_CHUNKS) = 1;
 
     // Copying buffer into chunk.
     memcpy(src_chunk_buf + chunk_user_data_offset, buf, buf_len_bytes);
@@ -367,7 +367,7 @@ EXTERN_C uint32_t __stdcall sc_bmx_send_buffer(
     assert(shared_memory_chunk::link_terminator != *src_chunk_index);    
 
     // Points to user data offset in chunk.
-    int32_t chunk_user_data_offset = *(int32_t*)(src_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_USER_DATA_OFFSET_IN_SOCKET_DATA) +
+    uint16_t chunk_user_data_offset = *(uint16_t*)(src_chunk_buf + starcounter::MixedCodeConstants::CHUNK_OFFSET_USER_DATA_OFFSET_IN_SOCKET_DATA) +
         starcounter::MixedCodeConstants::CHUNK_OFFSET_SOCKET_DATA;
 
     // Adding connection flags.
